src/main/workers/resources_worker/requirements.txt
nameko==2.11.0
psycopg2-binary==2.7.6.1
python-dotenv==0.9.1
PyMySQL
PyMySQL[rsa]